+++ frameworks/Java/helidon/benchmark_config.json
@@ -13,7 +13,7 @@
         "approach": "Realistic",
         "classification": "Micro",
         "database": "Postgres",
-        "framework": "Helidon",
+        "framework": "Helidon Reactive",
         "language": "Java",
         "flavor": "None",
         "orm": "Raw",
@@ -21,7 +21,30 @@
         "webserver": "None",
         "os": "Linux",
         "database_os": "Linux",
-        "display_name": "Helidon",
+        "display_name": "Helidon Reactive",
+        "notes": "",
+        "versus": "Netty"
+      },
+      "helidon-nima": {
+        "json_url": "/json",
+        "plaintext_url": "/plaintext",
+        "db_url": "/db",
+        "query_url": "/queries?queries=",
+        "update_url": "/updates?queries=",
+        "fortune_url": "/fortunes",
+        "port": 8080,
+        "approach": "Realistic",
+        "classification": "Micro",
+        "database": "Postgres",
+        "framework": "Helidon Nima",
+        "language": "Java",
+        "flavor": "None",
+        "orm": "Raw",
+        "platform": "Netty",
+        "webserver": "None",
+        "os": "Linux",
+        "database_os": "Linux",
+        "display_name": "Helidon Nima",
         "notes": "",
         "versus": "Netty"
       }
