+++ frameworks/Java/helidon/config.toml
@@ -17,3 +17,20 @@ orm = "Raw"
 platform = "Netty"
 webserver = "None"
 versus = "Netty"
+
+[nima]
+urls.plaintext = "/plaintext"
+urls.json = "/json"
+urls.db = "/db"
+urls.query = "/queries?queries="
+urls.update = "/updates?queries="
+urls.fortune = "/fortunes"
+approach = "Realistic"
+classification = "Micro"
+database = "Postgres"
+database_os = "Linux"
+os = "Linux"
+orm = "Raw"
+platform = "Nima"
+webserver = "None"
+versus = "Netty"
