+++ frameworks/Java/helidon/nima/src/main/resources/application.yaml
@@ -1,0 +1,37 @@
+#
+# Copyright (c) 2022 Oracle and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+server:
+  sockets:
+    - name: "@default"
+      host: "127.0.0.1"
+      port: 8080
+      backlog: 8192
+      receive-buffer-size: 64000
+      write-queue-length: 8192
+      # use the following to disable asynchronous
+      # write-queue-length: 0
+      connection-options:
+        read-timeout-seconds: 0
+        connect-timeout-seconds: 0
+        send-buffer-size: 64000
+        receive-buffer-size: 64000
+  connection-providers:
+    "http_1_1":
+      validate-headers: false
+      validate-path: false
+      recv-log: false
+      send-log: false
