+++ frameworks/Java/helidon/reactive/src/main/java/io/helidon/benchmark/reactive/models/Fortune.java
@@ -1,0 +1,19 @@
+package io.helidon.benchmark.reactive.models;
+
+public final class Fortune {
+    public int id;
+    public String message;
+
+    public Fortune(int id, String message) {
+        this.id = id;
+        this.message = message;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
