--- frameworks/Java/helidon/src/main/java/io/helidon/benchmark/models/DbRepository.java
@@ -1,14 +1,0 @@
-package io.helidon.benchmark.models;
-
-import java.util.List;
-
-import io.helidon.common.reactive.Single;
-
-public interface DbRepository {
-
-    Single<World> getWorld(int id);
-
-    Single<World> updateWorld(World world);
-
-    Single<List<Fortune>> getFortunes();
-}
