--- frameworks/Java/helidon/src/main/java/io/helidon/benchmark/models/Fortune.java
@@ -1,19 +1,0 @@
-package io.helidon.benchmark.models;
-
-public final class Fortune {
-    public int id;
-    public String message;
-
-    public Fortune(int id, String message) {
-        this.id = id;
-        this.message = message;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
